Move individual tests file content into vcTests.py
